14728 review fixes
