+++ netbox/core/tables/__init__.py
@@ -1,3 +1,4 @@
 from .config import *
 from .data import *
 from .jobs import *
+from .plugins import *
