+++ netbox/extras/tables/tables.py
@@ -5,7 +5,7 @@ from django.conf import settings
 from django.utils.translation import gettext_lazy as _
 
 from extras.models import *
-from netbox.tables import BaseTable, NetBoxTable, columns
+from netbox.tables import NetBoxTable, columns
 from .template_code import *
 
 __all__ = (
@@ -20,7 +20,6 @@ __all__ = (
     'ImageAttachmentTable',
     'JournalEntryTable',
     'ObjectChangeTable',
-    'PluginTable',
     'SavedFilterTable',
     'TaggedItemTable',
     'TagTable',
@@ -508,24 +507,3 @@ class JournalEntryTable(NetBoxTable):
         default_columns = (
             'pk', 'created', 'created_by', 'assigned_object_type', 'assigned_object', 'kind', 'comments'
         )
-
-
-class PluginTable(BaseTable):
-    verbose_name = tables.Column()
-    name = tables.Column()
-    author = tables.Column()
-    author_email = tables.Column()
-    description = tables.Column()
-    version = tables.Column()
-
-    class Meta:
-        empty_text = _('No plugins found')
-        fields = (
-            'verbose_name', 'name', 'author', 'author_email', 'description', 'version',
-        )
-        default_columns = (
-            'verbose_name', 'name', 'author', 'author_email', 'description', 'version',
-        )
-        attrs = {
-            'class': 'table table-hover object-list',
-        }
